Port x86 backend to 64-bit and drop 32-bit support
x86-64 port fixes (keep the VM booting to the end of bootstrap; the
known deterministic SIGSEGV at boot end is still present on x86-64):

- interpreter: SysV AMD64 arg passing for primitive calls (save the
  bytecode ptr in r12 before esi/rsi is clobbered, pop the caller-side
  argument scaffolding with addl(esp, slotSize)); fix D-I send return to
  pop the extra receiver word in the arg region.
- assembler/mapping: 64-bit-safe encodings on x86-64 - REX handling,
  RIP-relative disp32 for external references (&last_Delta_fp/sp) with
  relocations, movabs for oop loads, oop materialization via r10/r11 for
  movl/cmp forms that cannot embed a 64-bit immediate, pushl/popl as
  pushq/popq, incl/decl via FF /0,/1 (0x40+rd is a REX prefix), call_C
  SysV arg-register moves, store_check with 64-bit card-table pointers.
- stubRoutines: sign-extend the ic_info word to 64 bits before the flag
  shift (decode_ic_info) so NLR offsets decode correctly; nlr_home is
  intptr_t (was int, mismatched the extern declaration on LP64).
- primitives: smi_prims_gen 64-bit stack-slot offsets and 64-bit
  multiply-overflow detection.
- runtime: os_darwin waitForWithTimeout fix for async DLL waits.
- process: nlr_home/_nlr_home widened to intptr_t.
- build.unix/Makefile: select the assembler/mapping backend from the
  target architecture (-arch on macOS, uname -m on Linux) instead of the
  host compiler.

Deprecate 32-bit x86 entirely:

- DELTA_X86_32 removed wholesale (no longer defined, not even as 0);
  DELTA_X86_64 is now unconditionally 1 (BytesPerNativeWord = 8).
- vm/asm/assembler_x86.cpp: flattened all DELTA_X86_64 guards - the file
  is only ever compiled for the x86 backend, which is now always 64-bit,
  so the code is unconditionally x86-64 (dead 32-bit branches removed).
- test/assembler: fix the Makefile (its license block was a C-style
  comment, so the encoder tests never built); drop the 32-bit encoder
  target and 32-bit golden blocks; add the missing last_Delta_pc stub to
  the aarch64 test harness.

+++ vm/runtime/process.cpp
@@ -63,7 +63,7 @@ OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISE
 // the delta call. This is necessary for a consistent stack during the delta call.
 extern "C" char* C_frame_return_addr;
 extern "C" bool have_nlr_through_C;
-extern "C" int nlr_home;
+extern "C" intptr_t nlr_home;
 extern "C" int nlr_home_id;
 extern "C" oop nlr_result;
